requirements.txt
pandas == 2.2.2
numpy == 1.26.4
matplotlib == 3.9.0
scikit-learn == 1.5.0
Flask == 3.0.3
seaborn == 0.13.2
xgboost == 2.0.3
pymongo == 4.7.3
certifi == 2024.6.2
ipykernel == 6.29.4
imblearn == 0.0
PyYAML == 6.0.1
scipy == 1.13.1

-e .